requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openai==1.3.0
aiofiles==23.2.1
